+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/gui/AutomationPanel.java
@@ -26,6 +26,7 @@ import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.lang.System.Logger.Level;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -36,15 +37,9 @@ import javax.swing.JButton;
 import javax.swing.JFileChooser;
 import javax.swing.JOptionPane;
 import javax.swing.JScrollPane;
-import javax.swing.JTabbedPane;
-import javax.swing.JTextPane;
 import javax.swing.JToolBar;
 import javax.swing.SwingUtilities;
 import javax.swing.filechooser.FileFilter;
-import javax.swing.text.BadLocationException;
-import javax.swing.text.Style;
-import javax.swing.text.StyleConstants;
-import javax.swing.text.StyledDocument;
 import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.TreePath;
 import org.apache.logging.log4j.LogManager;
@@ -151,18 +146,12 @@ public class AutomationPanel extends AbstractPanel implements EventConsumer {
     private JButton jobDownButton;
     private JButton addJobButton;
     private JButton removeJobButton;
-    private JTabbedPane tabbedPane;
     private JButton addTestButton;
     private JButton removeTestButton;
     private JButton optionsButton;
     private JXTreeTable tree;
     private PlanTreeTableModel treeModel;
-    private JScrollPane outputScrollpane;
-    private JTextPane outputArea;
     private AutomationPlan currentPlan;
-    private Style styleError;
-    private Style styleWarning;
-    private Style styleInfo;
 
     public AutomationPanel(ExtensionAutomation ext) {
         this.ext = ext;
@@ -174,7 +163,7 @@ public class AutomationPanel extends AbstractPanel implements EventConsumer {
         this.setLayout(new GridBagLayout());
 
         this.add(this.getToolbar(), LayoutHelper.getGBC(0, 0, 1, 1.0));
-        this.add(this.getTabbedPane(), LayoutHelper.getGBC(0, 1, 1, 1.0, 1.0));
+        this.add(this.getPlanScrollpane(), LayoutHelper.getGBC(0, 1, 1, 1.0, 1.0));
 
         ZAP.getEventBus()
                 .registerConsumer(this, AutomationEventPublisher.getPublisher().getPublisherName());
@@ -606,7 +595,6 @@ public class AutomationPanel extends AbstractPanel implements EventConsumer {
         setCurrentPlan(plan);
         ext.registerPlan(currentPlan);
         AutomationProgress progress = currentPlan.getProgress();
-        this.getOutputArea().setText(listToStr(progress.getAllMessages()));
         if (progress.hasErrors()) {
             View.getSingleton()
                     .showWarningDialog(Constant.messages.getString("automation.panel.load.error"));
@@ -619,7 +607,6 @@ public class AutomationPanel extends AbstractPanel implements EventConsumer {
 
     public void setCurrentPlan(AutomationPlan plan) {
         currentPlan = plan;
-        getOutputArea().setText("");
         getTreeModel().setPlan(currentPlan);
         getRunPlanButton()
                 .setEnabled(
@@ -639,15 +626,6 @@ public class AutomationPanel extends AbstractPanel implements EventConsumer {
             return list;
         }
         return Collections.emptyList();
-    }
-
-    private JTabbedPane getTabbedPane() {
-        if (this.tabbedPane == null) {
-            this.tabbedPane = new JTabbedPane();
-            this.tabbedPane.addTab("Plan", this.getPlanScrollpane());
-            this.tabbedPane.addTab("Output", getOutputScrollpane());
-        }
-        return this.tabbedPane;
     }
 
     private JScrollPane getPlanScrollpane() {
@@ -712,30 +690,6 @@ public class AutomationPanel extends AbstractPanel implements EventConsumer {
         getRemoveTestButton().setEnabled(false);
     }
 
-    private JScrollPane getOutputScrollpane() {
-        if (outputScrollpane == null) {
-            outputScrollpane = new JScrollPane();
-            outputScrollpane.setViewportView(this.getOutputArea());
-        }
-        return outputScrollpane;
-    }
-
-    private JTextPane getOutputArea() {
-        if (outputArea == null) {
-            outputArea = new JTextPane();
-            outputArea.setEditable(false);
-            styleError = this.getOutputArea().addStyle("Error", null);
-            StyleConstants.setIcon(styleError, RED_BALL_ICON);
-
-            styleWarning = this.getOutputArea().addStyle("Warning", null);
-            StyleConstants.setIcon(styleWarning, ORANGE_BALL_ICON);
-
-            styleInfo = this.getOutputArea().addStyle("Info", null);
-            StyleConstants.setIcon(styleInfo, WHITE_BALL_ICON);
-        }
-        return outputArea;
-    }
-
     private PlanTreeTableModel getTreeModel() {
         if (treeModel == null) {
             treeModel = new PlanTreeTableModel(new DefaultMutableTreeNode("Plan"));
@@ -743,18 +697,21 @@ public class AutomationPanel extends AbstractPanel implements EventConsumer {
         return treeModel;
     }
 
-    private static String listToStr(List<String> list) {
-        return String.join("\n", list);
+    private void outputMessage(String message, Level level) {
+        ext.getView()
+                .getOutputPanel()
+                .append(
+                        String.format("%-9s%s\n", getI18nLogLevel(level), message),
+                        ext.getOutputSource().getName());
     }
 
-    private void outputMessage(String message, Style style) {
-        StyledDocument doc = this.getOutputArea().getStyledDocument();
-        try {
-            doc.insertString(doc.getLength(), " ", style);
-            doc.insertString(doc.getLength(), "  " + message + "\n", null);
-        } catch (BadLocationException e) {
-            // Ignore
-        }
+    private static String getI18nLogLevel(Level level) {
+        return switch (level) {
+            case ERROR -> Constant.messages.getString("automation.output.level.error");
+            case WARNING -> Constant.messages.getString("automation.output.level.warning");
+            case INFO -> Constant.messages.getString("automation.output.level.info");
+            default -> level.getName();
+        };
     }
 
     private AutomationPlan getPlan(Event event) {
@@ -821,7 +778,6 @@ public class AutomationPanel extends AbstractPanel implements EventConsumer {
                 updateSaveButton(event, true);
                 break;
             case AutomationEventPublisher.PLAN_STARTED:
-                this.getOutputArea().setText("");
                 break;
             case AutomationEventPublisher.PLAN_FINISHED:
                 this.getRunPlanButton().setEnabled(true);
@@ -829,15 +785,15 @@ public class AutomationPanel extends AbstractPanel implements EventConsumer {
                 break;
             case AutomationEventPublisher.PLAN_ERROR_MESSAGE:
                 outputMessage(
-                        event.getParameters().get(AutomationEventPublisher.MESSAGE), styleError);
+                        event.getParameters().get(AutomationEventPublisher.MESSAGE), Level.ERROR);
                 break;
             case AutomationEventPublisher.PLAN_WARNING_MESSAGE:
                 outputMessage(
-                        event.getParameters().get(AutomationEventPublisher.MESSAGE), styleWarning);
+                        event.getParameters().get(AutomationEventPublisher.MESSAGE), Level.WARNING);
                 break;
             case AutomationEventPublisher.PLAN_INFO_MESSAGE:
                 outputMessage(
-                        event.getParameters().get(AutomationEventPublisher.MESSAGE), styleInfo);
+                        event.getParameters().get(AutomationEventPublisher.MESSAGE), Level.INFO);
                 break;
             case AutomationEventPublisher.PLAN_ENV_CREATED:
                 getTreeModel().envChanged();